Add script to clean formatting issues in CSV files
Adds a script which fixes incorrect spacing or delimiters in some of
the structured fields that make up the card.csv file, including
Image URLs and Variations. Enforcing a normalized format will allow
easier parsing and reduce errors downstream. The script can easily
be extended to validate and clean other fields as well.

This script depends only on the python3 standard library, and should
be invoked from the root of the repository.