Add pre-commit hook for the clean script
This patch adds configuration for a pre-commit
(https://pre-commit.com/) hook to be run. Currently it runs only the
cleaning script, but this could also be used to run other
validators, code formatters, tests, etc. For example, you might have
a script that fails if any line in `card.csv` refers to a Type that
isn't in type.csv or a Set that isn't in set.csv.

Install pre-commit with `pip install pre-commit` and then enable the
hooks with `pre-commit install`. They will run when you run `git
commit`. If they make any changes, the commit will be aborted so you
can verify the changes, do a `git add`, and `git commit` again.