Add script to validate cross-file references
This patch adds another script, this one validates any references
between csv files. For example, card.csv has a column for a card's
Types. If a user wants to create a database table where each card is
linked to its Types, that user is likely to use type.csv to populate
that table. Accordingly, it is important to verify that this table
is consistent with the values that are actually used in the
card.csv.

This patch also updates the pre-commit config to allow this script
to be run automatically on `git commit`, and it can also be run in
CI. Instead of making corrections, this script simply fails with a
non-zero exit code (because we don't know if the value is a typo or
a missing entry, and we don't know if any other columns need to be
populated).